% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adtdc}
\alias{adtdc}
\title{Longitudinal time-dependent covariate data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 9813 rows and 4 columns.
}
\usage{
adtdc
}
\description{
This data set contains longitudinal time-dependent covariate data
on ECOG101 and LDH.
}
\details{
\describe{
\item{\code{SUBJID}}{subject ID}
\item{\code{PARAMCD}}{parameter code}
\item{\code{ADT}}{analysis date}
\item{\code{AVAL}}{covariate value}
}
}
\keyword{datasets}
