% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aml}
\alias{aml}
\title{Acute myelogenous leukemia survival data from the survival package}
\format{
An object of class \code{data.frame} with 23 rows and 3 columns.
}
\usage{
aml
}
\description{
Survival in patients with acute myelogenous leukemia.
\describe{
\item{\code{time}}{Survival or censoring time}
\item{\code{status}}{censoring status}
\item{\code{x}}{maintenance chemotherapy given or not}
}
}
\keyword{datasets}
