% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{heart}
\alias{heart}
\title{Stanford heart transplant data from the survival package}
\format{
An object of class \code{data.frame} with 172 rows and 8 columns.
}
\usage{
heart
}
\description{
Survival of patients on the waiting list for the Stanford heart
transplant program.
\describe{
\item{\code{start, stop, event}}{entry and exit time and status for
the time interval}
\item{\code{age}}{age-48 years}
\item{\code{year}}{year of acceptance (in years after Nov 1, 1967)}
\item{\code{surgery}}{prior bypass surgery 1=yes, 0=no}
\item{\code{transplant}}{received transplant 1=yes, 0=no}
\item{\code{id}}{patient id}
}
}
\keyword{datasets}
