% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ingots}
\alias{ingots}
\title{The binary data from Cox and Snell (1989, pp. 10-11).}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 25 rows and 4 columns.
}
\usage{
ingots
}
\description{
The dataset consits of the number of ingots not ready for rolling
and the number of ingots ready for rolling for a number of
combinations of heating time and soaking time.
}
\details{
\describe{
\item{\code{Heat}}{The heating time}
\item{\code{Soak}}{The soaking time}
\item{\code{NotReady}}{Response indicator, with a value 1 for units
not ready for rolling (event) and a value of 0 for units ready for
rolling (nonevent)}
\item{\code{Freq}}{The frequency of occurrence of each combination of
\code{Heat}, \code{Soak}, and \code{NotReady}}
}
}
\keyword{datasets}
