% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{liver}
\alias{liver}
\title{The liver data used in SAS PROC PHREG documentation examples.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 418 rows and 7 columns.
}
\usage{
liver
}
\description{
This data set contains information on 418 patients with primary biliary
cirrhosis.
}
\details{
\describe{
\item{\code{Time}}{The follow-up time in years from the time of
registration. The event could be liver transplantation, death,
or the end of the study, whichever came first}
\item{\code{Status}}{A censoring indicator, where a value of 1
indicating a death event, and 0 indicating a censored observation
(the patient survived past the observation time)}
\item{\code{Age}}{The patient's age in years}
\item{\code{Albumin}}{Serum albumin level in g/dl}
\item{\code{Bilirubin}}{Serum bilirubin level in mg/dl}
\item{\code{Edema}}{Edema status, where a value of 0 indicates no
edema, 0.5 indicates edema successfully treated with
diuretics, and 1 indicates edema despite diuretic therapy}
\item{\code{Protime}}{Prothrombin time in seconds}
}
}
\keyword{datasets}
