% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.assess_phregr.R
\name{plot.assess_phregr}
\alias{plot.assess_phregr}
\title{Plot method for assess_phregr objects}
\usage{
\method{plot}{assess_phregr}(x, nsim = 20, ...)
}
\arguments{
\item{x}{An object of class \code{assess_phregr}.}

\item{nsim}{The number of simulation samples used in the plot.}

\item{...}{Ensures that all arguments starting from "..." are named.}
}
\value{
A list of ggplot2 objects for the line plots, one for each
covariate.
}
\description{
Generate line plots of observed and simulated paths of
standardized score processes of a Cox model.
}
\author{
Kaifeng Lu, \email{kaifenglu@gmail.com}
}
\keyword{internal}
