% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ipe.R
\name{plot.ipe}
\alias{plot.ipe}
\title{Plot method for ipe objects}
\usage{
\method{plot}{ipe}(x, time_unit = "day", show_hr = TRUE, show_risk = TRUE, ...)
}
\arguments{
\item{x}{An object of class \code{ipe}.}

\item{time_unit}{The time unit used in the input data.
Options are "day" (default), "week", "month", or "year".}

\item{show_hr}{Logical; whether to show hazard ratio on the KM plot.
Default is TRUE.}

\item{show_risk}{Logical; whether to show number at risk table
below the KM plot. Default is TRUE.}

\item{...}{Ensures that all arguments starting from "..." are named.}
}
\value{
A list of ggplot2 objects: \code{p_res} for box plot for AFT model
deviance residuals, \code{p_kmstar} for KM plot for counterfactual
untreated survival times (i.e., potential outcomes if all subjects
were untreated), and \code{p_km} for KM plot for counterfactual unswitched
survival times.
}
\description{
Generate box plot for AFT model deviance residuals and
Kaplan-Meier (KM) plot for potential outcomes of an ipe object.
}
\author{
Kaifeng Lu, \email{kaifenglu@gmail.com}
}
\keyword{internal}
