% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.tsesimp.R
\name{plot.tsesimp}
\alias{plot.tsesimp}
\title{Plot method for tsesimp objects}
\usage{
\method{plot}{tsesimp}(x, time_unit = "day", show_hr = TRUE, show_risk = TRUE, ...)
}
\arguments{
\item{x}{An object of class \code{tsesimp}.}

\item{time_unit}{The time unit used in the input data.
Options are "day" (default), "week", "month", or "year".}

\item{show_hr}{Logical; whether to show hazard ratio on the KM plot.
Default is TRUE.}

\item{show_risk}{Logical; whether to show number at risk table
below the KM plot. Default is TRUE.}

\item{...}{Ensures that all arguments starting from "..." are named.}
}
\value{
A list of ggplot2 objects: \code{p_res} for box plot for AFT model
deviance residuals, \code{p_km_switch} for KM plot for time from secondary
baseline to switch, and \code{p_km} for KM plot for counterfactual
unswitched survival times.
}
\description{
Generate box plot for AFT model deviance residuals and
Kaplan-Meier (KM) plots for a tsesimp object.
}
\author{
Kaifeng Lu, \email{kaifenglu@gmail.com}
}
\keyword{internal}
