% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.assess_phregr.R
\name{print.assess_phregr}
\alias{print.assess_phregr}
\title{Print method for assess_phregr objects}
\usage{
\method{print}{assess_phregr}(x, ...)
}
\arguments{
\item{x}{An object of class \code{assess_phregr}.}

\item{...}{Ensures that all arguments starting from "..." are named.}
}
\value{
A printout from the fit of an assessment of proportional hazards
assumption of a Cox model.
}
\description{
Prints the concise information of an assess_phregr fit.
}
\author{
Kaifeng Lu, \email{kaifenglu@gmail.com}
}
\keyword{internal}
