% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.ipcw.R
\name{print.ipcw}
\alias{print.ipcw}
\title{Print method for ipcw objects}
\usage{
\method{print}{ipcw}(x, ...)
}
\arguments{
\item{x}{An object of class \code{ipcw}.}

\item{...}{Ensures that all arguments starting from "..." are named.}
}
\value{
A printout from the fit of an inverse-probability of censoring
weights model.
}
\description{
Prints the concise information of an ipcw fit.
}
\author{
Kaifeng Lu, \email{kaifenglu@gmail.com}
}
\keyword{internal}
