% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.ipe.R
\name{print.ipe}
\alias{print.ipe}
\title{Print method for ipe objects}
\usage{
\method{print}{ipe}(x, ...)
}
\arguments{
\item{x}{An object of class \code{ipe}.}

\item{...}{Ensures that all arguments starting from "..." are named.}
}
\value{
A printout from the fit of an iterative parameter estimation.
}
\description{
Prints the concise information of an ipe fit.
}
\author{
Kaifeng Lu, \email{kaifenglu@gmail.com}
}
\keyword{internal}
