% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.liferegr.R
\name{print.liferegr}
\alias{print.liferegr}
\title{Print liferegr Object}
\usage{
\method{print}{liferegr}(x, ...)
}
\arguments{
\item{x}{The liferegr object to print.}

\item{...}{Ensures that all arguments starting from "..." are named.}
}
\value{
A printout from the fit of an accelerated failue time model.
}
\description{
Prints the concise information of liferegr fit.
}
\author{
Kaifeng Lu, \email{kaifenglu@gmail.com}
}
\keyword{internal}
