% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.logisregr.R
\name{print.logisregr}
\alias{print.logisregr}
\title{Print logisregr Object}
\usage{
\method{print}{logisregr}(x, ...)
}
\arguments{
\item{x}{The logisregr object to print.}

\item{...}{Ensures that all arguments starting from "..." are named.}
}
\value{
A printout from the fit of a logistic regression model.
}
\description{
Prints the concise information of logisregr fit.
}
\author{
Kaifeng Lu, \email{kaifenglu@gmail.com}
}
\keyword{internal}
