% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.msm.R
\name{print.msm}
\alias{print.msm}
\title{Print method for msm objects}
\usage{
\method{print}{msm}(x, ...)
}
\arguments{
\item{x}{An object of class \code{msm}.}

\item{...}{Ensures that all arguments starting from "..." are named.}
}
\value{
A printout from the fit of a marginal structural model.
}
\description{
Prints the concise information of an msm fit.
}
\author{
Kaifeng Lu, \email{kaifenglu@gmail.com}
}
\keyword{internal}
