% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.phregr.R
\name{print.phregr}
\alias{print.phregr}
\title{Print phregr Object}
\usage{
\method{print}{phregr}(x, ...)
}
\arguments{
\item{x}{The phregr object to print.}

\item{...}{Ensures that all arguments starting from "..." are named.}
}
\value{
A printout from the fit of a Cox proportional hazards model.
}
\description{
Prints the concise information of phregr fit.
}
\author{
Kaifeng Lu, \email{kaifenglu@gmail.com}
}
\keyword{internal}
