% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.rpsftm.R
\name{print.rpsftm}
\alias{print.rpsftm}
\title{Print method for rpsftm objects}
\usage{
\method{print}{rpsftm}(x, ...)
}
\arguments{
\item{x}{An object of class \code{rpsftm}.}

\item{...}{Ensures that all arguments starting from "..." are named.}
}
\value{
A printout from the fit of a rank-preserving structural failure
time model.
}
\description{
Prints the concise information of a rpsftm fit.
}
\author{
Kaifeng Lu, \email{kaifenglu@gmail.com}
}
\keyword{internal}
