% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.ipcw.R
\name{print.summary.ipcw}
\alias{print.summary.ipcw}
\title{Print method for summary.ipcw objects}
\usage{
\method{print}{summary.ipcw}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{An object of class \code{summary.ipcw}.}

\item{digits}{The number of significant digits to print.}

\item{...}{Additional arguments passed to \code{print.data.frame}.}
}
\value{
The input object, invisibly.
}
\description{
Prints a detailed summary of an IPCW fit organized to
align with the reporting recommendations of NICE DSU TSD 24.
}
\keyword{internal}
