% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.ipe.R
\name{print.summary.ipe}
\alias{print.summary.ipe}
\title{Print method for summary.ipe objects}
\usage{
\method{print}{summary.ipe}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{An object of class \code{summary.ipe}.}

\item{digits}{The number of significant digits to print.}

\item{...}{Additional arguments passed to \code{print.data.frame}.}
}
\value{
The input object, invisibly.
}
\description{
Prints a detailed summary of an ipe fit.
}
\keyword{internal}
