% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.msm.R
\name{print.summary.msm}
\alias{print.summary.msm}
\title{Print method for summary.msm objects}
\usage{
\method{print}{summary.msm}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{An object of class \code{summary.msm}.}

\item{digits}{The number of significant digits to print.}

\item{...}{Additional arguments passed to \code{print.data.frame}.}
}
\value{
The input object, invisibly.
}
\description{
Prints a detailed summary of an MSM fit.
}
\keyword{internal}
