% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.rpsftm.R
\name{print.summary.rpsftm}
\alias{print.summary.rpsftm}
\title{Print method for summary.rpsftm objects}
\usage{
\method{print}{summary.rpsftm}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{An object of class \code{summary.rpsftm}.}

\item{digits}{The number of significant digits to print.}

\item{...}{Additional arguments passed to \code{print.data.frame}.}
}
\value{
The input object, invisibly.
}
\description{
Prints a detailed summary of a rpsftm fit.
}
\keyword{internal}
