% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.tsegest.R
\name{print.summary.tsegest}
\alias{print.summary.tsegest}
\title{Print method for summary.tsegest objects}
\usage{
\method{print}{summary.tsegest}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{An object of class \code{summary.tsegest}.}

\item{digits}{The number of significant digits to print.}

\item{...}{Additional arguments passed to \code{print.data.frame}.}
}
\value{
The input object, invisibly.
}
\description{
Prints a detailed summary of a two-stage g-estimation fit.
}
\keyword{internal}
