% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.tsesimp.R
\name{print.summary.tsesimp}
\alias{print.summary.tsesimp}
\title{Print method for summary.tsesimp objects}
\usage{
\method{print}{summary.tsesimp}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{An object of class \code{summary.tsesimp}.}

\item{digits}{The number of significant digits to print.}

\item{...}{Additional arguments passed to \code{print.data.frame}.}
}
\value{
The input object, invisibly.
}
\description{
Prints a detailed summary of a simple two-stage estimation fit.
}
\keyword{internal}
