% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.tsegest.R
\name{print.tsegest}
\alias{print.tsegest}
\title{Print method for tsegest objects}
\usage{
\method{print}{tsegest}(x, ...)
}
\arguments{
\item{x}{An object of class \code{tsegest}.}

\item{...}{Ensures that all arguments starting from "..." are named.}
}
\value{
A printout from the fit of two-stage estimation with g-estimation.
}
\description{
Prints the concise information of a tsegest fit.
}
\author{
Kaifeng Lu, \email{kaifenglu@gmail.com}
}
\keyword{internal}
