% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.tsesimp.R
\name{print.tsesimp}
\alias{print.tsesimp}
\title{Print method for tsesimp objects}
\usage{
\method{print}{tsesimp}(x, ...)
}
\arguments{
\item{x}{An object of class \code{tsesimp}.}

\item{...}{Ensures that all arguments starting from "..." are named.}
}
\value{
A printout from the fit of simple two-stage estimation.
}
\description{
Prints the concise information of a tsesimp fit.
}
\author{
Kaifeng Lu, \email{kaifenglu@gmail.com}
}
\keyword{internal}
