% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{process_cov}
\alias{process_cov}
\title{helper: process a base covariate specification and ensure design columns exist in df}
\usage{
process_cov(base_cov, df)
}
\description{
helper: process a base covariate specification and ensure design columns exist in df
}
\keyword{internal}
