% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sexagg}
\alias{sexagg}
\title{Urinary tract infection data from the logistf package}
\format{
An object of class \code{data.frame} with 36 rows and 9 columns.
}
\usage{
sexagg
}
\description{
This data set deals with urinary tract infection in sexually active
college women, along with covariate information on age an
contraceptive use. The variables are all binary and coded
in 1 (condition is present) and 0 (condition is absent).
}
\details{
\describe{
\item{\code{case}}{urinary tract infection, the study outcome variable}
\item{\code{age}}{>= 24 years}
\item{\code{dia}}{use of diaphragm}
\item{\code{oc}}{use of oral contraceptive}
\item{\code{vic}}{use of condom}
\item{\code{vicl}}{use of lubricated condom}
\item{\code{vis}}{use of spermicide}
}
}
\keyword{datasets}
