% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{six}
\alias{six}
\title{The repeated measures data from the "Six Cities" study of the health
effects of air pollution (Ware et al. 1984).}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 64 rows and 6 columns.
}
\usage{
six
}
\description{
The data analyzed are the 16 selected cases in Lipsitz et al. (1994).
The binary response is the wheezing status of 16 children at ages 9,
10, 11, and 12 years. A value of 1 of wheezing status indicates the
occurrence of wheezing. The explanatory variables city of residence,
age, and maternal smoking status at the particular age.
}
\details{
\describe{
\item{\code{case}}{case id}
\item{\code{city}}{city of residence}
\item{\code{age}}{age of the child}
\item{\code{smoke}}{maternal smoking status}
\item{\code{wheeze}}{wheezing status}
}
}
\keyword{datasets}
