% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.ipe.R
\name{summary.ipe}
\alias{summary.ipe}
\title{Summary method for ipe objects}
\usage{
\method{summary}{ipe}(object, ...)
}
\arguments{
\item{object}{An object of class \code{ipe}.}

\item{...}{Additional arguments passed to or from other methods.}
}
\value{
An object of class \code{summary.ipe} containing the analysis
population, reporting checklist, treatment effect estimates, outcome-model
parameter estimates, and bootstrap summary.
}
\description{
Summarizes reporting items and treatment effect estimates from
an iterative parameter estimation fit.
}
\author{
Kaifeng Lu, \email{kaifenglu@gmail.com}
}
\keyword{internal}
