% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.tsesimp.R
\name{summary.tsesimp}
\alias{summary.tsesimp}
\title{Summary method for tsesimp objects}
\usage{
\method{summary}{tsesimp}(object, ...)
}
\arguments{
\item{object}{An object of class \code{tsesimp}.}

\item{...}{Additional arguments passed to or from other methods.}
}
\value{
An object of class \code{summary.tsesimp} containing the analysis
population, reporting checklist, treatment effect estimates, AFT and
outcome-model parameter estimates, and bootstrap summary.
}
\description{
Summarizes reporting items and treatment effect estimates from
a simple two-stage estimation fit.
}
\author{
Kaifeng Lu, \email{kaifenglu@gmail.com}
}
\keyword{internal}
