% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tobin}
\alias{tobin}
\title{Tobin's tobit data from the survival package}
\format{
An object of class \code{data.frame} with 20 rows and 3 columns.
}
\usage{
tobin
}
\description{
Data from Tobin's original paper.
\describe{
\item{\code{durable}}{Durable goods purchase}
\item{\code{age}}{Age in years}
\item{\code{quant}}{Liquidity ratio (x 1000)}
}
}
\keyword{datasets}
